[SPARK-48734][PYTHON][TESTS] Separate local cluster test from test_artifact to test_artifact_localcluster
### What changes were proposed in this pull request?

This PR proposes to separate local cluster test from test_artifact to test_artifact_localcluster

### Why are the changes needed?

Separate local cluster test from test_artifact to test_artifact_localcluster. Both are reusing the tests but logically running different tests.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Manually tested via CI.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47118 from HyukjinKwon/SPARK-48734.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
